Refactor tournament management interfaces and classes
Updated `IRetrieveTournamentsView.cs` to define a public `IView` interface with comprehensive XML documentation for its methods. Enhanced `RetrieveTournamentsAdapter.cs` to implement the `IAdapter` interface, adding error handling and query execution methods, along with improved documentation. Modified `RetrieveTournamentsPresenter.cs` to include a public `Presenter` class with XML comments, clarifying its role in managing tournament presentation logic. Overall, these changes improve code readability, maintainability, and documentation.
